[WIP] Refactoring: Splitting CmcdController into CmcdModel & CmcdController
